Rename lib/ directory to src/
