+++ src/crosshair.js
