+++ src/hotbar.js
