+++ src/perf_hooks_replacement.js
